requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements_dev.txt requirements_dev.in
#
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
build==1.4.2
    # via pip-tools
click==8.3.1
    # via pip-tools
coverage==7.13.5
    # via -r requirements_dev.in
django==6.0.3
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.13.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
djangorestframework==3.17.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.29.0
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements_dev.in
faker==40.11.1
    # via
    #   -r requirements_dev.in
    #   factory-boy
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
packaging==26.0
    # via
    #   build
    #   pytest
    #   wheel
pillow==12.1.1
    # via -r requirements.in
pip-tools==7.5.3
    # via -r requirements_dev.in
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.11
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyjwt==2.12.1
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via -r requirements_dev.in
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.15.7
    # via -r requirements_dev.in
sqlparse==0.5.5
    # via django
uritemplate==4.2.0
    # via drf-spectacular
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
